Create upgrade.sh
Hi,

I was sick of doing the upgrade job on my onw. So i've created this upgrade script. It's tested on debian 7.5.
The script depends on the standardtools wget, unzip and mysqldump. 

To start the script go to your INSTLLATIONFOLDER and type ./scripts/upgrade.sh
Now the scripts makes the following steps

* At first you have to give the script the version number you want upgrade / downgrade to.
* It compares that version with the one of your installed Rudeltippen. Therefor it extracts the version out of the app/views/main.html 
* Then it stops the running app: play stop
* It creates the folder /backup, if it doesn't exist
* In this folder it creates another folder with an unique timestamp for every upgrade run.
* Then it copies the app, conf, logs, public, scripts and tmp folder in that folder.
* After that it makes a mysql dump of the database. Therefor it extracts the credentials out of the application.conf
* Then it trys to download the wanted version of rudeltippen and extracts the zip file in a unique update folder in the tmp folder
* After that it replaces the app, conf and public folder
* Then it copies the application.conf out of the backup back in the conf folder
* At least it removes the tmp files and starts the app: play start

I'am curious to know if this script works even fine on your installation!
Please give me feedback, if there are problems! If everythings fine and the script could be part of the package i start a pull request!

Best regards
schwindelbub